+++ metrics/coverage_high_water_mark
@@ -1,1 +1,1 @@
-93.9700
+93.9800
