+++ metrics/mypy_high_water_mark
@@ -1,1 +1,1 @@
-92.1400
+92.1500
